Avoid clearing isolated state in live tests
When testing live streaming, we don't create a new thread and we patch
the `Live` module to ensure that the thread locals are not cleared
afterwards.

We were however always calling
`ActiveSupport::IsolatedExecutionState.clear` so we were losing that
state in tests.

Additionally since https://github.com/rails/rails/pull/55247 calling
`ActiveSupport::IsolatedExecutionState.clear` is unsafe in test code,
because we are in an execution context (for the test) and the call wipes
that out.

This fix is not perfect as `share_with` does a shallow copy so changes
from within the test streaming "thread" can leak out - I think that's a
fundamental flaw in how the `Live` module and thread state interract.