Restore initial condition around setting of time attributes when deserializing a job
After #55661 was merged a bug was reported that the behavior changed slightly. Moving the if into a method resulted in the enqueued_at and scheduled_at being set to nil if no data was passed. Instead to restore the prior behavior it should skip setting the attr if the data is missing in the job_data.