Refactor BelongsToAssociation treatment of composite keys
Based on feedback:

```
First I think we should be able to get rid of if
reflection_fk.is_a?(Array) in favor of treating everything as an
Array(), even single-column PK/FKs.

And then perhaps we could semantically split it into two parts: one part
for the "replace" flow and the other one is "nullify" flow where record
is nil.
```