Optionally skip bundler-audit.
Running `bin/rails app:update` with Rails 8.1 adds bin/bundler-audit and config/bundler-audit.yml even if it's not in the Gemfile already. This checks whether bundler_audit is in the bundle and otherwise skips for the app generator.

Adds config option to skip bundler-audit in new applications.

Extends --minimal option to include bundler-audit

Updates changelog

Updates test assertions re: bundler-audit;
the #generate_test_dummy method skips bundle and bundler-audit, so it should not be expected to be present