hidde backtrace when parallel tests get interrupted
Right now if one presses CONTROL+C during the
parallel test run, one get a very long
backtrace that is very annoying.

The same problem was solved 11 years ago
for the non parallel tests here:

* https://github.com/minitest/minitest/commit/b6ec36d086ac28e47e4e2dc7ffd995f462543f94
* https://github.com/minitest/minitest/issues/503

The reason why the 11 years old fix does not
work for the parallel test is, that the rescue
happens before the shutdown method gets executed.
The shutdown method is where things gets executed.
In my opinion, this is wrong, but my guess is, that
it was done so, because doing it otherwise would
probably require changes in the Minitest gem itself.

NOTE: this is my first PR on the rails gem, so I apologize
if I made something wrong. Also, there is a big probability
that there is a better solution or that I miss some edge
case that I am missing because of lack of knowledge.