Indent private methods with two spaces consistently and add a note about it
I noticed that the tutorial had the most `private` methods have two extra spaces of indentation in almost all cases so I made this consistent.

I also added a note that this tutorial is following this convention. While I've seen it quite often, I know there seems to be competing [conventions](https://fabiokung.com/2010/04/05/ruby-indentation-for-access-modifiers-and-their-sections/) around this.