[Getting Started Tutorial] Add note on default string type for `email` column for Subscriber
When I was doing the tutorial, I noticed that the `email` column was not specified with a type yet it defaulted to a string. This confused me so I added a note for it with showing the generated migration file.

I tried looking for documentation on this, but didn't find anything. When I tried it in a few different projects, this did work. I did find the relevant code in `railties/lib/rails/generators/generated_attribute.rb` where
```ruby
def initialize(name, type = nil, index_type = false, attr_options = {})
  @name           = name
  @type           = type || :string
```

In the tutorial I don't want to link to the actual codebase, but I think there's a few more pages that I can add a note too.
* https://guides.rubyonrails.org/active_record_migrations.html
* https://guides.rubyonrails.org/command_line.html
* https://guides.rubyonrails.org/active_record_basics.html

Not sure if there's any other API specifc pages I can add a note to. If I do add these, should this be in this PR or a separate one?