Deprecate usage of custom ActiveJob serializers
The performance gains from https://github.com/rails/rails/pull/55583 meant that, when testing upgrades from 8.0 to 8.1, suddenly the app crashed and stopped evaluating due to the raised NotImplemented error in the parent class. This indexer isn't strictly required (there's a fallback), so it feels appropriate to do a proper deprecation cycle and then remove this behavior in a later version.