`ActionController::Parameters#deconstruct_keys` for pattern matching
Related to [#48429][]

Define `ActionController::Parameters#deconstruct_keys` in a `Hash`-like
way, so that `params` can integrate with [pattern matching][] (for
statements like `case … in`, `if … in`, and `params => { … }`).

To do so, declare the [#deconstruct_keys][] method in way that loops
over the specified keys and returns a
`ActiveSupport::HashWithIndifferentAccess` instance with only the
root-level keys. Any primitives will remain primitives, and any
`ActionController::Parameters` values below the first level of nesting
will remain `ActionController::Parameters` instances.

This integration can be useful for controller parameters (either at the
root level or nested within) that are to be passed to objects that don't
expect attributes or keyword arguments:

```ruby
if params in { search:, page: }
  Model.search(search).limit(page)
else
  …
end

case (value = params[:primitive_or_nested])
in String
  # do something with a String `value`…
in { nested_key: }
  # do something with `nested_key` or `value`
else
  # …
end
```

[#48429]: https://github.com/rails/rails/issues/48429
[pattern matching]: https://docs.ruby-lang.org/en/master/syntax/pattern_matching_rdoc.html
[#deconstruct_keys]: https://docs.ruby-lang.org/en/master/syntax/pattern_matching_rdoc.html#label-Matching+non-primitive+objects-3A+deconstruct+and+deconstruct_keys