Don't output deprecated message in newly generated applications
Since #55496, the depreciation message about `raise_on_open _redirects`
is shown in newly generated applications.

This is because the deprecations is shown when `raise_on_open_redirects`
is not null, but the default value of it is `false`.
https://github.com/rails/rails/blob/ec4337aae7d358e9d8ae3fc539347a2618b8fbe0/actionpack/lib/action_controller/railtie.rb#L110
https://github.com/rails/rails/blob/ec4337aae7d358e9d8ae3fc539347a2618b8fbe0/actionpack/lib/action_controller/railtie.rb#L15

In addition that, we set the value in `load_defaults`.
https://github.com/rails/rails/blob/ec4337aae7d358e9d8ae3fc539347a2618b8fbe0/railties/lib/rails/application/configuration.rb#L270

This PR removed config key and checked the key existence for the
condition of deprecation message.

Fixes #55772.