Store destroy dependent record message on reflection name
Currently we store the error on the `base` attribute which does
not allow to translate individual error messages per assocation.

Furthermore the record name was downcased not allowing to have
a capitalized record, which makes the error message hard to use
in places where the associated record is not an english word.

Fixes #53448