Add link-local IP ranges to RemoteIp default proxies
Link-local addresses (169.254.0.0/16 for IPv4 and fe80::/10 for IPv6)
should be treated as trusted proxies by default, similar to private
IP ranges, since they are non-routable and used for local network
communication.

Use of link-local IP addresses in tests were updated to use the 2001:db8::/32
netblock (RFC3849) to prevent conflict with the changes to the trusted
proxies.