Improve flow in Getting Started guide: create edit view before adding edit link
The guide previously instructed readers to add an Edit link to the show page
before creating the edit.html.erb view, which would result in broken links.

This commit reorders the sections to:
1. First introduce partials and create _form.html.erb
2. Update new.html.erb to use the partial
3. Create edit.html.erb using the partial
4. Then add the Edit link to show.html.erb

This ensures the edit view exists before we link to it, providing a smoother
tutorial experience for readers following along.