Refactor gcs_service public interface to bring more in line with s3_service
The s3_service exposes it's respective client and bucket attributes
publicly. I think this makes sense in the context of active storage, as
it allows users to rely on the same adapter & configs that the framework
is using for Activestorage without having to duplicate your own client
and ensure it matches the original, which can be modified at runtime, in
order to perform operations outside of the scope of ActiveStorage.

This commit serves to make the same true of the gcs_service, which
currently defined it's respective client & bucket as private method
definitions.

I did a bit of git snooping and didn't see any mention of this being an
intentional decision when first released.

Lint