Standardize event name formatting in `assert_event_reported`
This change adds `.inspect` to the event name in `assert_event_reported`'s
failure message, making it consistent with how `payload` and `tags` and
`assert_events_reported` are displayed.

`.inspect` is better and it helps distinguish between event names that
are strings vs symbols and is now consistent with the other fields.