[Getting Started] Use .positive? for more explicit positive checks rather than implicitly field itself
In the tutorial we use `product.inventory_count?` to check if the product is in stock. This is relying on implicit rails behavior that can be confusing and hard to reason about for newcomers.

This change uses `product.inventory_count.positive?` instead, which is more explicit and easier to understand. I also standardized on using it throughout the tutorial for consistency.