Make the warning about active storage redirect and proxy mode stronger
The existing documentation implied that the "hard to guess" URLs active storage
generates provides some sort of access control. This is not the case. Rather,
as they rely on ActiveRecord::SignedId, they're tamper proof and don't expose
the underlying id.

The security risk of using these URLs isn't that someone will guess them.
Rather, it is if the URLs are ever leaked, the files will be exposed.

This happened to us via Cloudflare's crawler hints feature:
https://www.tokyodev.com/articles/identifying-and-fixing-a-resume-visibility-mistake

But there is a myriad of other ways it could happen. E.g. anyone with access to
the request log has access to them, they might be exposed via certain analytics
tools, even if a user has access to the app revoked they'll still have access
to any files, etc.

Because of this, whenever any access control is needed, proxy and redirect mode
should not be used.

Co-authored-by: Petrik de Heus <petrik@deheus.net>