De-couple `@rails/actiontext/attachment_upload.js` from `Trix.Attachment`
`AttachmentUpload` property access
---

First, change the `AttachmentUpload` implementation to be less coupled
to Trix and [Trix.Attachment][] instances. Add an optional `file`
argument to its constructor.

When the `file` argument **is not** specified, it will be set to a
default value of `attachment.file`. In that case, the constructor
assumes that `attachment` is an instance of `Trix.Attachment`, and will
read a [File][] instance from its `.file` property. The `attachment`
argument is retained and set to an `.attachment` property. Subsequent
`direct-upload:`-prefixed events will continue to be dispatched with the
value of the `attachment` argument assigned to the event's
`detail.attachment` property. **This aims to maintain backwards
compatibility**.

When the `file` argument **is** specified and set to a [File][]
instance, it will be used throughout the rest of the implementation.
This enables future Action Text editors that are not Trix to provide
`File` instances directly in case their variety of "attachment" does not
have `.file` property.

Replace `AttachmentUpload` calls with events
---

Next, remove any calls to `Trix.Attachment.setUploadProgress` and
replace them with code in the `@rails/actiontext/index.js` module that
listens for corresponding `direct-upload:progress` event listeners.

Make `AttachmentUpload.start()` return a [Promise][]
---

Maintain the behavior that dispatches `direct-upload:error` and
`direct-upload:end` events. In addition to those hooks, editors can also
wait for the resolution of the [Promise][] instance returned by
`AttachmentUpload.start()`. The resolved value will include the `sgid`
and `url` attributes. Remove calls to `Trix.Attachment.setAttributes`,
and instead move the assignment of those attributes to the
`@rails/actiontext/index.js` module's event listeners.

[Promise]: https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Promise
[File]: https://developer.mozilla.org/en-US/docs/Web/API/File
[Trix.Attachment]: https://github.com/basecamp/trix/blob/v2.1.15/src/trix/models/attachment.js