Introduce SQLite3Adapter.resolve_path
A SQLite database may be configured with a relative path, an absolute
path, or a nonstandard `file:` URIs (possibly with params). When it is
a URI, it's non-trivial to determine the filesystem path to the
database.

This commit introduces a new singleton method on the connection
adapter to reliably resolve a filesystem path from the database
configuration.

It uses this method in the connection adapter's initializer to ensure
the database file's directory exists. Previously the adapter omitted
this check for database URIs.

It also introduces usage of this method in the `create` and `drop`
database tasks which also previously did not fully support database
URIs.