[Getting Started] Introduce flash earlier in the tutorial and give it more detail [ci skip]
We introduce flash quite late in the tutorial, section 16, but I think it can be useful to introduce it earlier since we can utilize it with the earlier CRUD actions. This way the reader of the tutorial gets some more debug information while running their scripts and have more than just a 422 http error to know a validation failed. Exposing the reader to the official flash API to handle notices felt quite worth it so I added some more explanation of how it works as well.