Fix passing both module: and shallow: to resources
This broke because the keyword arg change started always passing
`shallow:` through to this `scope` block. However, this caused an issue
because passing the `shallow` option caused it to _always_ override the
scope's current `shallow` value, instead of only doing that when
specified.

This commit fixes the behavior by only passing `shallow` to the `scope`
when it is specified, instead of unconditionally.

Co-authored-by: Marek Kasztelnik <mkasztelnik@gmail.com>