Update shard swap prohibition to be more granular
In #43485, shard swapping prohibition was introduced. It was
implemented as a boolean that applied to all databases (in the
isolated execution state).

I believe the spirit of the original feature was to prevent swapping
shards on a specific database, and I don't believe this feature took
into account the use case of a multi-database application.

This change scopes the shard swap prohibition to the "connection
specification", which allows an application to prohibit shard swapping
on a specific database while allowing it on all others.