Allow autosave association methods to be overwritten
If you try to override the `autosave` option by calling an association
method like `has_many(:some_existing_association, autosave: true)` in
a prepended module, it does not change the behavior of the underlying
class.

This is because even though you can override the reflection attached
to the class, the original reflection is still referenced within
blocks that define the methods related to autosaving associations.

The `add_autosave_association_callbacks` method are only defined when
an association is first built; or in the example I gave above, when an
association is rebuilt, so removing this check for method definition
should have a minimal performance impact.