Prefer Adapter#schema_cache to Model#schema_cache
Model#schema_cache has to lookup the current pool, but we already have a
connection/pool inside InsertAll, so this lookup is not needed.

This lookup currently happens 6 times inside InsertAll, and removing
them all saves about ~3.5% of time on the `/updates` TechEmpower
benchmark when using a SQLite memory database.