Fix ActiveSupport::Configurable deprecation
Before this commit, the constant could be referenced without a
deprecation warning:

```
$ ./tools/console
irb(main):001> ActiveSupport::Configurable
=> ActiveSupport::Configurable
irb(main):002>
```

This commit fixes the deprecation to fire when the file is required
(which should happen when the constant is referenced, since it is
autoloaded). It also removes the `eager_autoload` to prevent the
deprecation from warning when eager loading.

```
$ ./tools/console
irb(main):001> ActiveSupport::Configurable
DEPRECATION WARNING: ActiveSupport::Configurable is deprecated without
replacement, and will be removed in Rails 8.2.

You can emulate the previous behavior with `class_attribute`.
 (called from <top (required)> at (irb):1)
=> ActiveSupport::Configurable
irb(main):002>
```