[RF-Docs] [ci-skip]The Rails Command Line Guide  (#55612)
The goal of this PR is to update the Rails Command Line Guide. 

The existing guide was a bit sparse, so this diff has a decent amount of newly written content. This addresses feedback from the community (via Discord) that is noted in Basecamp task. Here are some of the highlights:

- reorganize the guide to add sections and subsections based on use case (instead of the command)
- new introduction and overview section
- expand the console section and add new dbconsole section
- added code snippets and why-is-this-command-useful to many commands
- add a new section for all of the commands in the db: namespace
- expand the testing section
- add a number of new commands that are new to Rails or were just missing from guides (e.g.bin/rails credentials:)

Co-authored-by: Harriet Oughton <73295547+OughtPuts@users.noreply.github.com>
Co-authored-by: Petrik de Heus <petrik@deheus.net>
Co-authored-by: Alex Kitchens <alexcameron98@gmail.com>