Refactors `ActiveRecord::Reflection::MacroReflection#_klass` inline rescue
Follow up #55949

The #55949 reverted #53969 because of the regression introduced by that PR, which broke resolution on nested models when a top-level module with the same name as the nested model exists. With that change the original issue that was raised in #53969 has been reverted as well.

The inline `rescue` could be mistaken for having the same functionality as a regular `rescue` block, and it is confusing to some people. This commit converts the inline `rescue` to a normal `begin / rescue` block, which makes it easier to understand and avoid potential issues in refactoring in the future.