[Getting Started] Add file paths to code blocks
I talked to some people who did the tutorial and found that they were
finding the blocks confusing since the filename was not always
mentioned, especially later in the tutorial. For someone familiar with
rails and path conventions its easy to find the way back so I didn't
notice this at first but I see how this can helped. After talking to
refaelfranca on the rails discord I decided to explore it.