Require 'rails' at the top of railltie files to ensure Rails is loaded first
This is consistent with what we do in Engines.

This is needed because we should not try to load any Rails framework
without requiring the top level file of that framework.

This is due to the fact that we avoid to cherry pick requires into
files when we can use shared requires in the top level file of each
framework.

Thus is important because 1db66c0 made a change to the railties gem to
follow the same convention we use in other frameworks, but it broke
some of the Railties because they are also entry points, so the
internal "active_support/rails" was never required.

Fixes #55970.