Enable `config.asset_host` to read from environment by default
This makes it so no code changes are needed in order to setup a CDN to
serve static assets.

If `ENV[CDN_HOST]` is not set, then `config.asset_host` will be set to
`nil`, meaning this commit does not change behavior.

Chose `ENV[CDN_HOST]` to be consistent with the [example][] used in the
Guides.

[example]: https://guides.rubyonrails.org/asset_pipeline.html#set-up-a-cdn-to-serve-static-assets