[Getting Started Guide] Add nesting for all routes snippets
As a first time reader to the tutorial, it might be confusing when to nest inside `Rails.application.routes.draw` DSL so I figured it's good to be explicit for this tutorial.

We already did explicitly show the nesting in some snippets of the tutorial so I'm making it consistent.