ActiveJob/serializers: Only index new serializers
Instead of reprocessing every serializer everytime a new serializer is added, just process the new one.
This has the intended side-effect of only issuing the deprecation warning for non-compliant serializers
once (when added) instead of on every subsequent call to add serializers. This also ensures that the
filepath in the deperecation warning does not point to the innocent compliant serializers.