[Getting Started] Add callout about automatic table matching in migrations [ci skip]
I noticed that the guides rely on automatic table matching in migrations, but for someone new to rails it's not obvious why this works and if they had a typo why the generated file didn't match.

I added a callout to the guides to explain how this works and since it's a nice rails convention over configuration feature it's good to highlight in intro tutorials.