Don't destroy ActiveStorage::Blob when its file deletion on the service fails
File deletion failure on the service did not restore the blob record which
results in a deleted DB record and a kept orphaned file in the storage.

GCS sometimes fails with Google::Cloud::UnavailableError. In such cases,
we want to keep the blob record intact so that we can retry deletion later.