Reset parallel test tables by deleting
When doing parallel testing with many databases, truncating all the
tables is very slow on MySQL. Deleting from the tables is much faster
even if they contain a reasonable amount of data.

Running on a schema with about 200 empty tables and 32 parallel
databases, deleting from all tables takes about 0.3 seconds, while
truncating takes 18 seconds. Even with 1000 rows to delete in each
table deleting is about 2.7x faster than truncating.

On PostgreSQL, truncating the tables is slower than deleting if they
have small amounts of data which is typical for test databases. Both
methods are very quick anyway.

There's no difference between the two methods for SQLite, as the adapter
uses DELETE for truncation anyway.

Functionally DELETE and TRUNCATE are identical on PostgreSQL, but on
MySQL TRUNCATE will reset AUTO_INCREMENT counters while DELETE will not.

Tests should not be relying on them being reset and since this is only
for parallel tests where test orders are going to be randomised anyway
so the IDs are unlikely to be consistent between test runs anyway.

By switching to DELETE all three databases will behave identically
and not reset the ID sequence counters.

Timings for resetting tables with TRUNCATE vs DELETE, based on 32
databases with 200 tables each, with 0, 100 and 1000 rows per table:
```
  ┌──────────────┬─────────┬──────────┬───────┬───────┬───────┬─────────┐
  │   Database   │  Rows/  │  Method  │  Min  │  Avg  │  Max  │ Winner  │
  │              │  Table  │          │ (sec) │ (sec) │ (sec) │         │
  ├──────────────┼─────────┼──────────┼───────┼───────┼───────┼─────────┤
  │              │         │ TRUNCATE │ 0.269 │ 0.535 │ 0.578 │         │
  │ PostgreSQL   │    0    │──────────┼───────┼───────┼───────┼─────────┤
  │              │         │ DELETE   │ 0.038 │ 0.051 │ 0.067 │  10.6x  │
  ├──────────────┼─────────┼──────────┼───────┼───────┼───────┼─────────┤
  │              │         │ TRUNCATE │ 0.174 │ 0.323 │ 0.408 │         │
  │ PostgreSQL   │   100   │──────────┼───────┼───────┼───────┼─────────┤
  │              │         │ DELETE   │ 0.104 │ 0.132 │ 0.150 │  2.45x  │
  ├──────────────┼─────────┼──────────┼───────┼───────┼───────┼─────────┤
  │              │         │ TRUNCATE │ 0.511 │ 0.578 │ 0.604 │  2.48x  │
  │ PostgreSQL   │  1000   │──────────┼───────┼───────┼───────┼─────────┤
  │              │         │ DELETE   │ 1.372 │ 1.434 │ 1.568 │         │
  ├──────────────┼─────────┼──────────┼───────┼───────┼───────┼─────────┤
  │              │         │ TRUNCATE │17.485 │18.413 │18.978 │         │
  │ MySQL        │    0    │──────────┼───────┼───────┼───────┼─────────┤
  │              │         │ DELETE   │ 0.311 │ 0.341 │ 0.356 │  53.6x  │
  ├──────────────┼─────────┼──────────┼───────┼───────┼───────┼─────────┤
  │              │         │ TRUNCATE │16.892 │18.281 │18.845 │         │
  │ MySQL        │   100   │──────────┼───────┼───────┼───────┼─────────┤
  │              │         │ DELETE   │ 0.534 │ 1.406 │ 2.003 │  13.0x  │
  ├──────────────┼─────────┼──────────┼───────┼───────┼───────┼─────────┤
  │              │         │ TRUNCATE │16.723 │17.939 │18.544 │         │
  │ MySQL        │  1000   │──────────┼───────┼───────┼───────┼─────────┤
  │              │         │ DELETE   │ 3.789 │ 6.635 │ 8.400 │  2.70x  │
  └──────────────┴─────────┴──────────┴───────┴───────┴───────┴─────────┘
```

If there's much more data in the tables then TRUNCATE may become
faster, but in that case the tests are likely to be slow as they'll need
to generate the data.

This PR also removed the SKIP_TEST_DATABASE_TRUNCATE environment
variable - we no longer truncate so the name doesn't make sense and
deleting should be fast enough to not need to skip.