[Getting Started] Rewrite Continiouus Integration section to use `bin/ci`
With Rails 8.1, we have introduced `bin/ci` as the starter to CI pipelines.
Currently, the guide focused on GitHub Actions as the starter which is tied
to the specific platform.

This change rewrites the Continuous Integration section in Getting Started guide
to use `bin/ci` as the main example, making it platform-agnostic, emphasizing
local execution and introducing the recent capabilities of `bin/ci`.