Add `:algorithim` option to `has_secure_password`
This is a new take on #48993, with a branch up to speed with recent `main`
and without the base class or the `available_password_algorithms` registry like Hash.

libraries/apps can just bring an object that responds to the required methods and `has_secure_password`
will just work with the custom implementation.