Add class option `primary_key_default` for migration and model generator
The purpose of this commit is to add a way to set the uuidv7 function (introduced in PostgreSQL 18) as the default value in migrations.
With the existing generator, it was possible to change the primary_key_type, but not the default value itself — resulting in gen_random_uuid() being automatically selected.
To address this, a new option primary_key_default has been added, allowing developers to choose any UUID generation function they prefer.