Make structured events for Active Record work
Currently, structured events for Active Record doesn't emit on v8.1.
This is because `active_record/structured_event_subscriber` isn't loaded.
I guess we just forgot to require it in #55690.

In `main` branch, this was already fixed by #55900.
I think we should fix this on 8.1 because we mentioned structured
events for Active Record on CHANGELOG.
https://github.com/rails/rails/blob/8-1-stable/activerecord/CHANGELOG.md#rails-810-october-22-2025