Fix devcontainer volumes for PostgreSQL 18
The current configuration prints a warning when using the version 18
image:

```
Important Change: the PGDATA environment variable of the image was
changed to be version specific in PostgreSQL 18 and above. For 18 it is
/var/lib/postgresql/18/docker. Later versions will replace 18 with their
respective major version (e.g., /var/lib/postgresql/19/docker for
PostgreSQL 19.x). The defined VOLUME was changed in 18 and above to
/var/lib/postgresql. Mounts and volumes should be targeted at the
updated location. This will allow users upgrading between PostgreSQL
major releases to use the faster --link when running pg_upgrade and
mounting /var/lib/postgresql.
```

The recommendation is now to mount the entire `/var/lib/postgresql`
folder so that it will hold multiple, versioned data dirs inside it.