Fix DelegateClass defining ... signatures
Previously, the `signature` variable was only getting set in the first
two branches, but not the third. This caused methods that _should_ have
been defined with triple dot delegation to instead have no arguments
defined.

This commit fixes the issue by pulling the `signature =` assignment out
of the condition so that it always happens.