PostgreSQLAdapter: fully reset the connection when reconnecting
Ref: https://github.com/rails/rails/issues/51780#issuecomment-3571653983

Postgres connections require a lot of associated state in `@type_map` and
other values. The configuration process is very long and contain many
pieces.

If it's interupted by something like `Timeout.timeout`, all bets
are off as to what state we're in, and we may end up in a finally
broken state where the connection report being active but is unusable.

To ensure that even such broken connection won't remain in such state
forever, this commit de-optimize `PostgreSQLAdapter#reconnect`.

We no longer try to reset and re-use the connection, we do a full
reconnect and reconfigure.

Big thanks to Shayon Mukherjee for the repro.