Schedule a single job per Blob for creating variants
If an ActiveStorage attachment has multiple preprocessed variants, a job
gets created for each variant, resulting in a lot of jobs when multiple
blobs are attached with multiple variants.

For a previewable blob, these jobs get batched into a single
PreviewImageJob that processes the preview and schedules jobs for all
preprocessed variants.

For non-previewable blobs we can do the same: schedule a single job that
schedules jobs for all preprocessed variants.

Co-authored-by: Tom Rossi <tom@higherpixels.com>